+++ src/main/java/com/ahr/xtaldb/UserInfo.java
@@ -4,7 +4,7 @@ import java.util.LinkedList;
 
 public class UserInfo {
     User user=null;
-    int money=0;
+    float money=0;
     LinkedList<Product> products=null;
     LinkedList<Product> wishlist=null;
     UserInfo(){}
